\name{ecodata2}
\alias{ecodata2}
\title{Environmental data set to exemplify the use of the functions contained in \emph{NonParRolCor} 
}
\description{
The data set \code{\link{ecodata2}} contains three columns, the first one (named ''Years'') is the time (years from 1700 to 1936, yearly resolution), the second column (named ''TSI'') are reconstructions of total solar irradiance (Lean 2000) and the third column the first component principal (PC1) of the reconstructed Atlantic Bluefin Tuna (BFT) captures (Ganzedo et al. 2016, Polanco-Martínez et al. 2018). 
}
\usage{data(ecodata2)
}
\format{One file in ASCII format containing three columns and 237 rows, columns are separated by spaces. 
}
\source{
Lean, J. (2000). Evolution of the Sun's spectral irradiance since the Maunder Minimum. Geophysical Research Letters, 27(16), 2425-2428. <URL: \doi{10.1029/2000GL000043}>. Lean Web TSI data set: <URL: \url{https://www.ncei.noaa.gov/access/paleo-search/study/5788/}>. \cr 

Ganzedo, U., Polanco-Martínez, J. M., Caballero-Alfonso, A. M., Faria, S. H., Li, J., Castro-Hernández, J. J. (2016). Climate effects on historic bluefin tuna captures in the Gibraltar Strait and Western Mediterranean. Journal of Marine Systems, 158, 84-92. \cr 
<URL: \doi{10.1016/j.jmarsys.2016.02.002}>. \cr

Polanco-Martínez, J. M., Caballero-Alfonso, A. M., Ganzedo, U., Castro-Hernández, J. J. (2018). A reconstructed database of historic bluefin tuna captures in the Gibraltar Strait and Western Mediterranean. Data in Brief, 16, 206-210. <URL: \doi{10.1016/j.dib.2017.11.028}>. 
}
\keyword{dataset3}
