\name{syntheticdata}
\alias{syntheticdata}
\title{Synthetic data set to exemplify the use of the functions contained in \emph{NonParRolCor} 
}
\description{
The data set \code{\link{syntheticdata}} contains three columns: the first one are the ``times'' (from 1 to 500) (named ``Times''), the second (named ``X'') and the third (named ``Y'') columns were generated by a bi-variate AR1 process with similar autocorrelation coefficients of 0.25. We generate two correlated bi-variate AR1 time series with positive (direct) correlation (0.85) for the first 250 elements and with negative (inverse) correlation (-0.85) for the last 250 elements (Polanco-Martínez and López-Martínez 2021). 
}
\usage{data(syntheticdata)
}
\format{One file in ASCII format containing three columns and 500 rows, columns are separated by spaces. 
}
\source{
Author's own production (Josué M. Polanco-Martínez) based on: mpiktas (<URL: \url{https://stats.stackexchange.com/users/2116/mpiktas/}>). How to simulate two correlated AR(1) time series?, Cross Validated. (2013). <URL: \url{https://stats.stackexchange.com/q/71831/}> (version: 2013-10-03).\cr

Polanco-Martínez, J. M. and López-Martínez, J.M. (2021). A non-parametric method to test the statistical significance in rolling window correlations, and applications to ecological time series, Ecological Informatics 60, 101379. <URL: \doi{10.1016/j.ecoinf.2021.101379}>. 
} 
\keyword{synthetic_dataset}
